backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
web3==7.6.0
eth-account==0.13.4
requests==2.31.0
gunicorn==21.2.0

# Optional - for PostgreSQL in production
# Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
requests==2.31.0
gunicorn==21.2.0
